requirements.txt
# Core data science libraries (Windows-compatible versions)
pandas>=1.5.0
numpy>=1.21.0
scipy>=1.8.0

# Web scraping and APIs
requests>=2.28.0
yfinance>=0.2.0
beautifulsoup4>=4.11.0
sec-edgar-api>=1.1.0

# Machine Learning
scikit-learn>=1.1.0
joblib>=1.2.0
xgboost>=1.6.0
statsmodels>=0.13.0
arch>=5.3.0

# Deep Learning (CPU versions for compatibility)
torch>=1.12.0
torchvision>=0.13.0

# Graph Neural Networks
torch-geometric>=2.4.0

# Natural Language Processing
nltk>=3.8
vaderSentiment>=3.3.2

# Visualization
matplotlib>=3.5.0
plotly>=5.10.0
seaborn>=0.11.0

# Web interface
streamlit>=1.25.0

# Explainability
shap>=0.48.0
lime>=0.2.0

# Experiment tracking
mlflow>=2.9.0

# Data processing
openpyxl>=3.0.10
python-dateutil>=2.8.2

# Configuration and utilities
python-dotenv>=0.20.0
tqdm>=4.64.0

# News API integration
gnews>=0.3.7
